requirements.txt
Django>3.2,<3.3
pillow
django-multi-captcha-admin
django-robots
django-simple-captcha
django-summernote
django-taggit
django-debug-toolbar
django_extensions
python-decouple
djangorestframework
markdown
django-filter
# coreapi
drf-yasg[validation]
djangorestframework-simplejwt